requirements.txt
nextcord==2.0.0
dadjokes==1.3.2
nekosbest==1.1.8
PyNaCl==1.5.0
python-dotenv==0.19.2
randfacts==0.15.1

